Autonomous Deep Research Analyst/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deep-research-analyst"
version = "0.1.0"
description = "Autonomous multi-agent deep research system built with LangGraph — web search, synthesis, and PDF/HTML report generation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "typing-extensions>=4.5",
    "langchain-core>=0.2",
    "langchain-groq>=0.1",
    "langchain-openai>=0.1",
    "langgraph>=0.2",
    "tavily-python>=0.3",
    "markdown>=3.5",
    "matplotlib>=3.8",
    "pandas>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "ipykernel>=6.0",
]

[project.scripts]
deep-research = "src.cli.interactive:interactive_loop"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
